refactor(CustomDateTimeFormatterTest): main 패키지와 동일하도록 /util 아래로 위치 이동
