+++ src/test/java/roomescape/util/CustomDateTimeFormatterTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,7 +6,6 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.util.CustomDateTimeFormatter;
 
 class CustomDateTimeFormatterTest {
